+++ gDRutils/inst/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Change to v.1.3.15
+- Add tests for convert_se_ref_assay_to_dt
+
 # Change to v.1.3.14
 - Add helpers for colData/rowData refinement